re-factoring code
