+++ geometry.py
@@ -5,779 +5,12 @@
 import numpy as np
 from copy import deepcopy
 
-import myMath
 import grid
+from body import Body
+from region import Region
+from transformation import RotDefi, Transformation
+from scorings import Usrbin
 
-StringPrec=1.0E-14 # prec in determining integer values
-MaxLineLength=132
-LineHeader="%13s"%("")
-# FREE format:
-cut0sFREE="0"*4
-expDigitsFREE=22
-numDigitsFREE=expDigitsFREE-7
-# FIXED format:
-cut0sFIXED="0"*2
-expDigitsFIXED=10
-numDigitsFIXED=expDigitsFIXED-7
-
-class GeoObject():
-    '''
-    A very basic object, implementing simply a name and a comment.
-    This object is supposed to collect all the stuff concerning
-      handling of names and modification of comments of instances
-      of the other classes;
-    '''
-
-    def __init__(self,myName="",myComment=""):
-        self.name=myName
-        self.comment=myComment
-
-    def rename(self,newName,lNotify=True):
-        'change instance name'
-        if (lNotify):
-            self.tailMe("* NAME CHANGE: FROM %s TO %s"%(self.name,newName))
-        self.name=newName
-
-    def headMe(self,myString):
-        'head a string to the comment'
-        if (len(self.comment)>0):
-            self.comment=myString+"\n"+self.comment
-        else:
-            self.comment=myString
-            
-    def tailMe(self,myString):
-        'tail a string to the comment'
-        if (len(self.comment)>0):
-            self.comment=self.comment+"\n"+myString
-        else:
-            self.comment=myString
-                
-    def echoComm(self):
-        myBuf=""
-        if (len(self.comment)>0):
-            myBuf=self.comment+"\n"
-        return myBuf
-
-    def echoName(self):
-        return self.name
-            
-class Body(GeoObject):
-    '''
-    - supported bodies: planes, spheres, and TRCs, nothing else for the time being;
-    - body definition on ONE line only, and always starts at col 1;
-    - support only for preceding comments, NO comments headed by "!" or after
-      body definition;
-    - no support for $start_* cards;
-    '''
-
-    def __init__(self,myName="",myComment=""):
-        '''
-        default body is a FLUKA PLA
-        '''
-        GeoObject.__init__(self,myName=myName,myComment=myComment)
-        self.bType="PLA"
-        self.P=np.zeros(3)
-        self.V=np.array([0.0,0.0,1.0])
-        self.Rs=np.zeros(2)
-        self.lIsRotatable=True
-        self.TransformName=None
-
-    def echo(self,lFree=True,maxLen=MaxLineLength,header=LineHeader,lMultiLine=True):
-        '''take into account comment'''
-        myStr="%-3s %8s"%(self.bType,self.echoName())
-        if (self.bType=="PLA"):
-            myFloats=list(self.V)+list(self.P)
-        elif (self.bType=="YZP"): 
-            myFloats=self.P[0]
-        elif (self.bType=="XZP"):
-            myFloats=self.P[1]
-        elif (self.bType=="XYP"):
-            myFloats=self.P[2]
-        elif (self.bType=="SPH"):
-            myFloats=list(self.P)+[self.Rs[0]]
-        elif (self.bType=="TRC"):
-            myFloats=list(self.P)+list(self.V)+list(self.Rs)
-        elif (self.bType=="RCC"):
-            myFloats=list(self.P)+list(self.V)+[self.Rs[0]]
-        elif (self.bType=="XCC"):
-            myFloats=list(self.P[1:])+[self.Rs[0]]
-        elif (self.bType=="YCC"):
-            myFloats=[self.P[2],self.P[0],self.Rs[0]]
-        elif (self.bType=="ZCC"):
-            myFloats=list(self.P[0:-1])+[self.Rs[0]]
-        else:
-            print("body %s NOT supported yet!"%(self.bType))
-            exit(1)
-        myStrings=[myStr]+echoFloats(myFloats,lFree=lFree)
-        myStr=assembleLine(myStrings,maxLen=maxLen,header=header,lMultiLine=lMultiLine)
-        if (self.TransformName is not None):
-            myStr="$Start_transform -%s\n%s\n$end_transform"%(self.TransformName,myStr)
-        return GeoObject.echoComm(self)+myStr
-
-    def isRotatable(self):
-        return self.lIsRotatable
-
-    def linkTransformName(self,Tname=None):
-        if (Tname is not None or Tname!=""):
-            self.TransformName=Tname
-            
-    def retTransformName(self):
-        return self.TransformName
-
-    @staticmethod
-    def fromBuf(tmpLines):
-        newBody=Body()
-        for tmpLine in tmpLines.splitlines():
-            data=tmpLine.split()
-            if (data[0]=="PLA"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.V=np.array(data[2:5]).astype(float)
-                newBody.P=np.array(data[5:8]).astype(float)
-            elif (data[0]=="YZP"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P[0]=float(data[2])
-                newBody.lIsRotatable=False
-            elif (data[0]=="XZP"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P[1]=float(data[2])
-                newBody.lIsRotatable=False
-            elif (data[0]=="XYP"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P[2]=float(data[2])
-                newBody.lIsRotatable=False
-            elif (data[0]=="SPH"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P=np.array(data[2:5]).astype(float)
-                newBody.Rs[0]=float(data[5])
-            elif (data[0]=="TRC"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P=np.array(data[2:5]).astype(float)
-                newBody.V=np.array(data[5:8]).astype(float)
-                newBody.Rs=np.array(data[8:]).astype(float)
-            elif (data[0]=="RCC"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P=np.array(data[2:5]).astype(float)
-                newBody.V=np.array(data[5:8]).astype(float)
-                newBody.Rs[0]=float(data[8])
-            elif (data[0]=="XCC"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P[1:]=np.array(data[2:4]).astype(float)
-                newBody.Rs[0]=float(data[4])
-                newBody.lIsRotatable=False
-            elif (data[0]=="YCC"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P[0]=data[3]; newBody.P[2]=data[2]
-                newBody.Rs[0]=float(data[4])
-                newBody.lIsRotatable=False
-            elif (data[0]=="ZCC"):
-                newBody.bType=data[0]
-                newBody.rename(data[1],lNotify=False)
-                newBody.P[0:-1]=np.array(data[2:4]).astype(float)
-                newBody.Rs[0]=float(data[4])
-                newBody.lIsRotatable=False
-            elif (tmpLine.startswith("*")):
-                newBody.tailMe(tmpLine)
-            else:
-                print("body %s NOT supported yet!"%(data[0]))
-                exit(1)
-        return newBody
-
-    def makeRotatable(self,lDebug=False,infL=1000.0):
-        if (not self.isRotatable()):
-           myStr=GeoObject.echoComm(self)+"%-3s %8s"%(self.bType,self.echoName())
-           if (self.bType=="YZP"):
-               if (lDebug):
-                   print("...converting body %s into a PLA"%(myStr))
-               self.bType="PLA"
-               self.V=np.array([1.0,0.0,0.0])
-           elif (self.bType=="XZP"):
-               if (lDebug):
-                   print("...converting body %s into a PLA"%(myStr))
-               self.bType="PLA"
-               self.V=np.array([0.0,1.0,0.0])
-           elif (self.bType=="XYP"):
-               if (lDebug):
-                   print("...converting body %s into a PLA"%(myStr))
-               self.bType="PLA"
-               self.V=np.array([0.0,0.0,1.0])
-           elif (self.bType=="XCC"):
-               if (lDebug):
-                   print("...converting body %s into an RCC"%(myStr))
-               self.bType="RCC"
-               self.P[0]=-infL
-               self.V=np.array([2*infL,0.0,0.0])
-           elif (self.bType=="YCC"):
-               if (lDebug):
-                   print("...converting body %s into an RCC"%(myStr))
-               self.bType="RCC"
-               self.P[1]=-infL
-               self.V=np.array([0.0,2*infL,0.0])
-           elif (self.bType=="ZCC"):
-               if (lDebug):
-                   print("...converting body %s into an RCC"%(myStr))
-               self.bType="RCC"
-               self.P[2]=-infL
-               self.V=np.array([0.0,0.0,2*infL])
-           else:
-               print("cannot make body %s rotatable!"%(myStr))
-               exit(1)
-           self.lIsRotatable=True
-
-    def traslate(self,dd=None):
-        if (dd is not None):
-            self.P=self.P+dd
-        else:
-            print("please provide me with a 3D traslation array!")
-            exit(1)
-                
-    def rotate(self,myMat=None,myTheta=None,myAxis=3,lDegs=True,lDebug=False):
-        if (myMat is not None):
-            self.P=myMat.mulArr(self.P,lDebug=lDebug)
-            self.V=myMat.mulArr(self.V,lDebug=lDebug)
-        elif (myTheta is not None):
-            if (myAxis is None):
-                print("you must specify a rotation axis together with an angle!")
-                exit(1)
-            else:
-                myMat=myMath.RotMat(myAng=myTheta,myAxis=myAxis,lDegs=lDegs,lDebug=lDebug)
-                self.rotate(myMat=myMat,myTheta=None,myAxis=myAxis,lDegs=lDegs,lDebug=lDebug)
-
-    def resize(self,newL,lDebug=False):
-        if (self.bType=="RCC"):
-           if (lDebug):
-               myStr=GeoObject.echoComm(self)+"%-3s %8s"%(self.bType,self.echoName())
-               print("...resizing body %s to L=%g"%(myStr,newL))
-           meanPoint=self.P+self.V/2.
-           versor=self.V/np.linalg.norm(self.V)
-           self.V=versor*newL
-           self.P=meanPoint-self.V/2.
-                
-class Region(GeoObject):
-    '''
-    - no parsing/splitting of zones;
-    - support only for preceding comments or commented lines between region
-      definition lines, NO comments headed by !
-    - a region definition always starts at column 1;
-    - no check of length of lines for region definition;
-    - no support for LATTICE;
-    - ASSIGNMA cards: only 1:1 material:region assignment, NO material
-      assignment for decay radiation simulation, NO magnetic/electric fields;
-    '''
-
-    def __init__(self,myName="",myComment=""):
-        GeoObject.__init__(self,myName=myName,myComment=myComment)
-        self.neigh=5
-        self.definition=""
-        self.material="BLACKHOLE"
-        # additional fields
-        self.initCont()
-
-    def initCont(self,rCont=0,rCent=np.array([0.0,0.0,0.0]),rMaxLen=0.0):
-        self.rCont=rCont # containment indicator (1 per region):
-                         # -1: region to be contained into/sized by another one
-                         #  0: regular region (neither contains nor it is contained)
-                         #  1: cell region (it contains another region)
-        self.rCent=rCent # central point of one or more region (3D array)
-        self.rMaxLen=rMaxLen # max length from hive cell
-
-    @staticmethod
-    def fromBuf(myBuffer):
-        newReg=Region()
-        lHeadParsed=False
-        for tmpLine in myBuffer.splitlines():
-            if (not lHeadParsed):
-                if (tmpLine.startswith("*")):
-                    newReg.tailMe(tmpLine)
-                else:
-                    data=tmpLine.split()
-                    newReg.rename(data[0],lNotify=False)
-                    newReg.neigh=float(data[1])
-                    newReg.definition=tmpLine
-                    # remove region name and number of neighbour regions from
-                    #       definition
-                    newReg.definition=newReg.definition.replace(data[0],"",1)
-                    newReg.definition=newReg.definition.replace(data[1],"",1)
-                    # remove heading/trailing empty spaces
-                    newReg.definition=newReg.definition.strip() 
-                    lHeadParsed=True
-            else:
-                newReg.definition=newReg.definition+"\n"+tmpLine
-        return newReg
-        
-    def BodyNameReplaceInDef(self,oldNames,newNames):
-        'query-replace body names in region definition'
-        for oName,nName in zip(oldNames,newNames):
-            self.definition=self.definition.replace(oName,nName)
-
-    def assignMat(self,myMaterial):
-        self.material=myMaterial
-
-    def addZone(self,myDef,nSpace=16):
-        if (self.isNonEmpty()):
-            if ( '|' not in self.definition ):
-                # if more than a zone, be sure that the first zone has the "|" preceeding char,
-                #    just for the sake of proper alignment
-                self.definition="| %s"%(self.definition)
-            self.definition=self.definition+'\n'+" "*nSpace+'| %s'%(myDef)
-        else:
-            self.definition=myDef
-
-    def isNonEmpty(self):
-        return len(self.definition)>0
-        
-    def echo(self,lMat=False):
-        '''take into account comment'''
-        if (lMat):
-            # echo ASSIGNMA card
-            return GeoObject.echoComm(self)+"ASSIGNMA  %10s%10s" % ( self.material, self.echoName() )
-        else:
-            return GeoObject.echoComm(self)+"%-8s   %4d %s" % \
-                ( self.echoName(), self.neigh, self.definition )
-
-    def retBodiesInDef(self):
-        'return list of body names in region definition'
-        bodiesInDef=[]
-        if (self.isNonEmpty):
-            for myLine in self.definition.splitlines():
-                if (myLine.startswith("*")): continue # comment line
-                tmpLine=myLine.replace("|"," ")
-                tmpLine=tmpLine.replace("+"," ")
-                tmpLine=tmpLine.replace("-"," ")
-                currBodies=tmpLine.strip().split()
-                for currBody in currBodies:
-                    tmpBody=currBody.strip()
-                    if (tmpBody not in bodiesInDef):
-                        bodiesInDef.append(tmpBody)
-        else:
-            print("Region %s is empty!"%(self.echoName()))
-            exit(1)
-        return bodiesInDef
-
-    def merge(self,newReg,spacing=" "*16):
-        # warn user in comment
-        self.tailMe("* --> merged with region %s <--"%(newReg.echoName()))
-        # merge comments
-        if (len(newReg.comment)>0):
-            self.tailMe(newReg.comment)
-        # merge definitions
-        mergedDef=""
-        newRegZones=newReg.definition.split("|")
-        myRegZones=self.definition.split("|")
-        lFirst=True
-        for ii,myRegZone in enumerate(myRegZones,1):
-            sMyRegZone=myRegZone.strip()
-            if (len(sMyRegZone)>0):
-                for jj,newRegZone in enumerate(newRegZones,1):
-                    sNewRegZone=newRegZone.strip()
-                    if (len(sNewRegZone)>0):
-                        tmpComment="* merging zone %s:%d into %s:%d"%(\
-                                newReg.echoName(),jj,self.echoName(),ii)
-                        if (lFirst):
-                            self.tailMe(tmpComment)
-                            mergedDef="| %s %s"%(sMyRegZone,sNewRegZone)
-                            lFirst=False
-                        else:
-                            mergedDef=mergedDef+"\n%s"%(tmpComment)+"\n%s| %s %s"%(spacing,sMyRegZone,sNewRegZone)
-        self.definition=mergedDef
-        # merge neighbours
-        self.neigh=self.neigh+newReg.neigh
-        # remove any sign of merge labelling
-        self.initCont()
-
-class RotDefi(GeoObject):
-    '''
-    Class handling ROT-DEFI cards.
-    Please have a look also to the Transformation() class: the ROT-DEFI
-       class simply stores infos of a specific rotation/traslation;
-    - echo in FREE format uses an empty space as field separator;
-    - parsing in FREE format expects an empty space as field separator;
-    '''
-
-    def __init__(self,myAx=3,myPhi=0.0,myTh=0.0,myDD=np.zeros(3),myComment=""):
-        # mainly for comments, name NOT used!
-        GeoObject.__init__(self,myComment=myComment)
-        self.axis=myAx           # FLUKA coding: 1=x, 2=y, 3=z
-        self.phi=myPhi           # polar angle [degrees] (FLUKA units)
-        self.theta=myTh          # azimuth angle [degrees] (FLUKA units)
-        self.DD=myDD             # translation [cm] (x,y,z components)
-
-    def echo(self,lFree=True,myID=0,myName="",\
-             maxLen=MaxLineLength,header=LineHeader,lMultiLine=False):
-        '''echo in FREE format uses an empty space as field separator'''
-        if (myID<=0):
-            print("...cannot dump a ROT-DEFI card without an index!")
-            exit(1)
-        if (len(myName)==0):
-            print("...cannot dump a ROT-DEFI card without a name!")
-            exit(1)
-        if (myID<100):
-            myIDecho=myID
-            if (self.phi!=0.0 or self.theta!=0.0):
-                myIDecho=myIDecho+self.axis*100
-        else:
-            myIDecho=myID*1000
-            if (self.phi!=0.0 or self.theta!=0.0):
-                myIDecho=myIDecho+self.axis
-        myFloats=[myIDecho,self.phi,self.theta]+list(self.DD)
-        myStrings=["ROT-DEFI  "]+echoFloats(myFloats,lFree=lFree)+[" %-10s"%(myName)]
-        return GeoObject.echoComm(self)+ \
-            assembleLine(myStrings,maxLen=maxLen,header=header,lMultiLine=lMultiLine)
-
-    @staticmethod
-    def fromBuf(myBuffer,lFree=True):
-        '''parsing in FREE format expects an empty space as field separator'''
-        newRotDefi=RotDefi()
-        for tmpLine in myBuffer.splitlines():
-            if (tmpLine.startswith("*")):
-                newRotDefi.tailMe(tmpLine.strip())
-            else:
-                # parse data
-                if (lFree):
-                    data=tmpLine.split()
-                    pID=data[1]
-                    phi=data[2]
-                    tht=data[3]
-                    DD=data[4:7]
-                    myName=data[7]
-                else:
-                    pID=tmpLine[10:20].strip()
-                    phi=tmpLine[20:30].strip()
-                    tht=tmpLine[30:40].strip()
-                    DD=[tmpLine[40:50].strip(),tmpLine[50:60].strip(),tmpLine[60:70].strip()]
-                    myName=tmpLine[70:80].strip()
-                # store data
-                pID=float(pID)
-                if (pID>1000):
-                    myID=int(pID/1000)
-                    myAxis=int(pID%1000)
-                else:
-                    myAxis=int(pID/100)
-                    myID=int(pID%100)
-                if (myAxis>0):
-                    newRotDefi.axis=myAxis
-                newRotDefi.phi=float(phi)
-                newRotDefi.theta=float(tht)
-                newRotDefi.DD=np.array([float(myDD) for myDD in DD])
-        return newRotDefi, myID, myName
-
-class Transformation(GeoObject):
-    '''
-    This is the actual class to be used when dealing with ROT-DEFI cards.
-    The class implements an array of ROT-DEFI cards - hence, at least one card.
-    '''
-
-    def __init__(self,myID=0,myName=""):
-        GeoObject.__init__(self,myName=myName) # mainly for names, comments NOT used!
-        self.RotDefis=[] # list of ROT-DEFI cards
-        self.ID=myID
-        
-    def __len__(self):
-        return len(self.RotDefis)
-
-    def __iter__(self):
-        self.current_index=0
-        return self
-    
-    def __next__(self):
-        '''
-        from https://blog.finxter.com/python-__iter__-magic-method/
-        '''
-        if self.current_index < len(self):
-            x = self.RotDefis[self.current_index]
-            self.current_index += 1
-            return x
-        raise StopIteration
-
-    # override methods of GeoObject, since comments are of the single
-    #   ROT-DEFI cards
-    def headMe(self,myString):
-        self.RotDefis[0].headMe(myString)
-    def tailMe(self,myString):
-        self.RotDefis[0].tailMe(myString)
-    def echoComm(self):
-        myComm=""
-        if (len(self.RotDefis)>0):
-            myComm=self.RotDefis[0].echoComm()
-        return myComm
-
-    def AddRotDefi(self,newRotDefi,iAdd=-1,lMoveComment=True):
-        if (iAdd==-1 or iAdd==len(self.RotDefis)):
-            # append ROT-DEFI card to list of ROT-DEFIs
-            self.RotDefis.append(newRotDefi)
-        elif (iAdd==0):
-            # head ROT-DEFI card to list of ROT-DEFIs
-            if (lMoveComment):
-                myComm=self.echoComm().strip()
-                if (len(myComm)>0):
-                    # take care of main comment
-                    newRotDefi.headMe(myComm)
-                    self.RotDefis[0].comment=""
-            self.RotDefis=[newRotDefi]+self.RotDefis
-        else:
-            # insert ROT-DEFI card at specific position in the list of ROT-DEFIs
-            self.RotDefis=self.RotDefis[0:iAdd]+[newRotDefi]+self.RotDefis[iAdd:]
-    def AddRotDefis(self,newRotDefis,iAdd=-1,lMoveComment=True):
-        for myRotDefi in newRotDefis:
-            self.AddRotDefi(myRotDefi,iAdd=iAdd,lMoveComment=lMoveComment)
-
-    @staticmethod
-    def fromBuf(self,myBuffer,lFree=True):
-        newTransf=Transformation()
-        myBuf=""
-        for tmpLine in myBuffer.splitlines():
-            if (tmpLine.startswith("*")):
-                if (len(myBuf)==0):
-                    myBuf=tmpLine
-                else:
-                    myBuf=myBuf+"\n"+tmpLine
-            else:
-                tmpRotDefi=RotDefi()
-                myID,myName=tmpRotDefi.fromBuf(myBuf)
-                if (myID!=0):
-                    if (newTransf.ID==0):
-                        # first ROT-DEFI card of transformation
-                        newTransf.ID=myID
-                    elif(newTransf.ID!=myID):
-                        print("...cannot add ROT-DEFI with ID=%d to tranformation ID=%d!"%(myID,newTransf.ID))
-                        exit(1)
-                if (len(myName)>0):
-                    if (len(newTransf.echoName())==0):
-                        # first ROT-DEFI card of transformation
-                        newTransf.rename(myName,lNotify=False)
-                    elif (newTransf.echoName()!=myName):
-                        print("...cannot add ROT-DEFI named %s to tranformation named %s !"%(myName,newTransf.echoName()))
-                        exit(1)
-                newTransf.AddRotDefi(tmpRotDefi)
-                myBuf=""
-        return newTransf
-
-    def echo(self,lFree=True):
-        buf=""
-        for myRotDefi in self.RotDefis:
-            if (len(buf)>0):
-                buf=buf+"\n"
-            buf=buf+myRotDefi.echo(myID=self.ID,myName=self.name)
-        return buf
-
-class Usrbin(GeoObject):
-    '''
-    A very basic class for handling USRBIN cards.
-    For the time being, there is support only for:
-    - the scoring is not really interpreted: the FLUKA definition is kept in
-      memory, and manipulations are performed on-the-fly;
-    - parsing/echoing, ALWAYS in FIXED format;
-    - always 2 cards to fully define scoring;
-    - a single ROTPRBIN card per transformation (1:1 mapping between ROTPRBIN and
-      USRBIN cards); the mapping to the transformation is ALWAYS name-based;
-      the ROTPRBIN card always PRECEEDs the respective USRBIN card;
-    - NO comments between USRBIN cards defining the same scoring detector;
-    - manipulation of unit, extremes and numbers of bins; nevertheless,
-      NO check is performed between the requested axis for manipulation and
-      the type of mesh;
-    '''
-    def __init__(self,myName="",myComment=""):
-        GeoObject.__init__(self,myName=myName,myComment=myComment)
-        self.definition=[] # array of strings storing the lines defining the
-                           #   USRBIN. NB: USRBIN tags, & char and bin name
-                           #   are NOT stored.
-        self.TransfName=""
-
-    def echo(self):
-        '''take into account comment'''
-        tmpBuf=GeoObject.echoComm(self)
-        if (len(self.TransfName)>0):
-            tmpBuf=tmpBuf+"%-10s%10s%10s%10s%10s\n"\
-                %("ROTPRBIN","",self.TransfName,"",self.echoName())
-        for myDef,mySdum,myEoL in zip(self.definition,[self.echoName(),"&"],["\n",""]):
-            tmpBuf=tmpBuf+"%-10s%60s%-10s%-s"%("USRBIN",myDef,mySdum,myEoL)
-        return tmpBuf
-                        
-    @staticmethod
-    def fromBuf(myBuffer):
-        newUsrBin=Usrbin()
-        for myLine in myBuffer.split("\n"):
-            myLine=myLine.strip()
-            if (myLine.startswith("*")):
-                newUsrBin.tailMe(myLine)
-            elif (myLine.startswith("ROTPRBIN")):
-                newUsrBin.assignTrasf(myLine[20:30].strip())
-                if (len(newUsrBin.TransfName)==0):
-                    print("...something wrong when parsing USRBIN cards: no ROT-DEFI card name!")
-                    exit(1)
-            else:
-                newUsrBin.definition.append(myLine[10:70])
-                if (len(newUsrBin.definition)==1):
-                    # from first line defining USRBIN, get USRBIN name
-                    newUsrBin.rename(myLine[70:].strip(),lNotify=False)
-        if (len(newUsrBin.definition)!=2):
-            print("...something wrong when parsing USRBIN cards: got %d lines!"\
-                  %(len(newUsrBin.definition)))
-            exit(1)
-        if (len(newUsrBin.name)==0):
-            print("...something wrong when parsing USRBIN cards: no name!")
-            exit(1)
-        return newUsrBin
-
-    def getUnit(self):
-        return float(self.definition[0][20:30])
-    def setUnit(self,myUnit,lFree=False):
-        if (self.getUnit()<0):
-            self.definition[0]=self.definition[0][ 0:20]+echoFloats(-abs(myUnit),lFree=lFree)[0]+ \
-                               self.definition[0][30:]
-        else:
-            self.definition[0]=self.definition[0][ 0:20]+echoFloats( abs(myUnit),lFree=lFree)[0]+ \
-                               self.definition[0][30:]
-
-    def isSpecialBinning(self):
-        binType=int(abs(float(self.definition[0][0:10]))+1E-4)
-        return binType==2.0 or binType==12.0 or binType==8.0 or binType==18.0
-
-    def getNbins(self,axes="all"):
-        nBins=None
-        if (not self.isSpecialBinning() ):
-            if (isinstance(axes,list)):
-                nBins=1
-                for myAx in axes:
-                    nBins=nBins*self.getNbins(axes=myAx)
-            elif (isinstance(axes,str)):
-                if (axes.upper()=="ALL"):
-                    nBins=self.getNbins(axes=[1,2,3])
-                elif (axes.upper()=="X"):
-                    nBins=self.getNbins(axes=[1])
-                elif (axes.upper()=="Y"):
-                    nBins=self.getNbins(axes=[2])
-                elif (axes.upper()=="Z"):
-                    nBins=self.getNbins(axes=[3])
-                else:
-                    print("cannot get number of bins on ax specified as %s!"%(axes))
-                    exit(1)
-            elif (1<=axes and axes<=3):
-                nBins=int(abs(float(self.definition[1][30+(axes-1)*10:40+(axes-1)*10]))+1E-4)
-            else:
-                print("cannot get number of bins on ax %d [1:3]!"%(axes))
-                exit(1)
-        return nBins
-    def setNbins(self,nBins=[1],axes=[3],lFree=False):
-        if (isinstance(axes,float) or isinstance(axes,int)): axes=[axes]
-        elif (isinstance(axes,str)):
-            if (axes.upper()=="ALL"):
-                axes=[1,2,3]
-            else:
-                print("cannot set nBins to ax %s!"%(axes))
-                exit(1)
-        if (isinstance(nBins,float) or isinstance(nBins,int)): nBins=[nBins]
-        if (len(axes)!=len(nBins)):
-            print("len(axes)!=len(nBins)! axes=",axes,"; nBins=",nBins)
-            exit(1)
-        for myN,myAx in zip(nBins,axes):
-            if (myAx<3):
-                self.definition[1]=self.definition[1][0:30+10*(myAx-1)]+\
-                   echoFloats(myN,lFree=lFree)[0]+self.definition[1][30+10*myAx:]
-            elif (myAx>0):
-                self.definition[1]=self.definition[1][0:30+10*(myAx-1)]+\
-                   echoFloats(myN,lFree=lFree)[0]
-            else:
-                print("cannot set %d bins to axis %d!"%(myN,myAx))
-                exit(1)
-
-    def getExtremes(self,axes=3):
-        myMin=None; myMax=None
-        if (isinstance(axes,list)):
-            myMin=[]; myMax=[]
-            for myAx in axes:
-                tmpMin,tmpMax=self.getExtremes(axes=myAx)
-                myMin.append(tmpMin); myMax.append(tmpMax)
-        elif (isinstance(axes,float) or isinstance(axes,int)):
-            if (1<=axes and axes<=2):
-                myMax=float(self.definition[0][30+(axes-1)*10:30+axes*10])
-            elif (axes==3):
-                myMax=float(self.definition[0][30+(axes-1)*10:          ])
-            else:
-                print("cannot get extremes of bin on axis %d!"%(axes))
-                exit(1)
-            myMin=float(self.definition[1][ (axes-1)*10:axes*10])
-        elif (isinstance(axes,str)):
-            if (axes.upper()=="X"):
-                myMin,myMax=self.getExtremes(self,axes=1)
-            elif (axes.upper()=="Y"):
-                myMin,myMax=self.getExtremes(self,axes=2)
-            elif (axes.upper()=="Z"):
-                myMin,myMax=self.getExtremes(self,axes=3)
-            else:
-                print("cannot get extremes of bin on axis %s!"%(axes))
-                exit(1)
-        return myMin,myMax
-    def setExtremes(self,myMin,myMax,axes=3,lFree=False):
-        if (isinstance(myMin,list) and isinstance(myMax,list) and isinstance(axes,list)):
-            if (len(myMin)!=len(myMax) or len(myMin)!=len(axes)):
-                print("cannot set min=",myMin,",max=",myMax,"axes",axes)
-                exit(1)
-            for tMin,tMax,tAx in zip(myMin,myMax,axes):
-                self.setExtremes(tMin,tMax,axes=tAx)
-        elif (not isinstance(myMin,list) and not isinstance(myMax,list) and not isinstance(axes,list)):
-            myMinStr=echoFloats(myMin,lFree=lFree)[0]
-            myMaxStr=echoFloats(myMax,lFree=lFree)[0]
-            if (isinstance(axes,float) or isinstance(axes,int)):
-                if (1<=axes and axes<=2):
-                    self.definition[0]=self.definition[0][:30+(axes-1)*10]+\
-                                       myMaxStr+\
-                                       self.definition[0][30+axes*10:]
-                elif (axes==3):
-                    self.definition[0]=self.definition[0][:30+(axes-1)*10]+\
-                                       myMaxStr
-                else:
-                    print("cannot set extremes of bin on axis %d!"%(axes))
-                    exit(1)
-                self.definition[1]=self.definition[1][:(axes-1)*10]+\
-                                   myMinStr+\
-                                   self.definition[1][axes*10:]
-            elif (isinstance(axes,str)):
-                if (axes.upper()=="X"):
-                    self.setExtremes(self,myMin,myMax,axes=1)
-                elif (axes.upper()=="Y"):
-                    self.setExtremes(self,myMin,myMax,axes=2)
-                elif (axes.upper()=="Z"):
-                    self.setExtremes(self,myMin,myMax,axes=3)
-                else:
-                    print("cannot set extremes of bin on axis %s!"%(axes))
-                    exit(1)
-        else:
-            print("mixed arrays!")
-            exit(1)
-
-    def resize(self,newL,axis=3):
-        if (axis!=3):
-            print("For the time being, it is possible to re-size USRBINs only along the Z-axis!")
-            exit(1)
-        currMin,currMax=self.getExtremes(axes=axis); currNbins=self.getNbins(axes=axis)
-        currDelta=currMax-currMin; currStep=currDelta*1.0/currNbins
-        currMean=(currMax+currMin)*0.5
-        # actually resize
-        newNbins=newL*1.0/currDelta*currNbins
-        if (newNbins%1<0.5):
-            newNbins=int(newNbins)
-        else:
-            newNbins=int(newNbins)+1
-        newMin=currMean-0.5*newNbins*currStep
-        newMax=currMean+0.5*newNbins*currStep
-        self.setExtremes(newMin,newMax,axes=axis)
-        self.setNbins(nBins=newNbins,axes=axis)
-
-    def assignTrasf(self,trasfName):
-        if (len(trasfName)>0):
-            self.TransfName=trasfName
-    def returnTrasf(self):
-        return self.TransfName
-    
 class Geometry():
     '''
     - name-based FLUKA geometry defition;
@@ -1813,79 +1046,6 @@ def acquireGeometries(fileNames,geoNames=None,lMakeRotatable=False):
         print("...done;")
         
     return myGeos
-
-def echoFloats(myFloats,expDigits=None,numDigits=None,prec=StringPrec,cut0s=None,lFree=True):
-    if (lFree):
-        if (expDigits is None): expDigits=expDigitsFREE
-        if (numDigits is None): numDigits=numDigitsFREE
-        if (cut0s is None): cut0s=cut0sFREE
-    else:
-        if (expDigits is None): expDigits=expDigitsFIXED
-        if (numDigits is None): numDigits=numDigitsFIXED
-        if (cut0s is None): cut0s=cut0sFIXED
-    numFmt="%% %d.%dE"%(expDigits,numDigits)
-    if (isinstance(myFloats,list)):
-        floats2convert=myFloats
-    else:
-        floats2convert=[myFloats]
-    myStrs=[]
-    for float2convert in floats2convert:
-        if (abs(float2convert)%1<prec): # integer
-            myStr="% .1f"%(float2convert)
-        else: # actual floating point number
-            myStr=numFmt%(float2convert)
-            data=myStr.split("E")
-            mantissa=data[0]; exponent=data[1]
-            if (cut0s is not None): # truncate numbers if a series of "0" is found
-                if (mantissa.find(cut0s)>-1):
-                    mantissa=mantissa[:mantissa.index(cut0s)]
-            if (exponent=="+00"):
-                exponent=None
-            elif (exponent=="+01"):
-                exponent=None
-                mantissa=mantissa.replace(".","")
-                mantissa=mantissa[0:3]+"."+mantissa[3:]
-            elif (exponent=="+02"):
-                exponent=None
-                mantissa=mantissa.replace(".","")
-                mantissa=mantissa[0:4]+"."+mantissa[4:]
-            elif (exponent=="-01"):
-                exponent=None
-                mantissa=mantissa.replace(".","")
-                mantissa=mantissa[0]+"0."+mantissa[1:]
-            elif (exponent=="-02"):
-                exponent=None
-                mantissa=mantissa.replace(".","")
-                mantissa=mantissa[0]+"0.0"+mantissa[1:]
-            if (mantissa=="-0.0"): mantissa=" 0.0"
-            if (exponent is None):
-                myStr=mantissa
-            else:
-                myStr="%sE%s"%(mantissa,exponent)
-        if (not lFree):
-            if (len(myStr)<expDigits):
-                myStr=" "*(expDigits-len(myStr))+myStr
-        myStrs.append(myStr)
-    return myStrs
-
-def assembleLine(myStrings,maxLen=MaxLineLength,header=LineHeader,lMultiLine=True):
-    if (len(myStrings)==0):
-        finStr=""
-    else:
-        finStr=myStrings[0]; curLen=len(finStr)
-        for myStr in myStrings[1:]:
-            if (curLen+len(myStr)+1>=maxLen):
-                if (lMultiLine):
-                    finStr=finStr+"\n"+header+myStr
-                    curLen=len(header+myStr)
-                else:
-                    print("Cannot prepare a line longer than %d chars!"%(maxLen))
-                    print("%s %s"%(finStr,myStr))
-                    exit(1)
-            else:
-                finStr=finStr+" "+myStr
-                curLen=curLen+len(myStr)+1
-    return finStr
 
 if (__name__=="__main__"):
     lDebug=False
